+++ src/main/java/me/drex/worldmanager/extractor/RarArchiveExtractor.java
@@ -6,22 +6,25 @@ import com.github.junrar.exception.UnsupportedRarV5Exception;
 import com.github.junrar.rarfile.FileHeader;
 import com.mojang.brigadier.exceptions.CommandSyntaxException;
 import me.drex.worldmanager.save.WorldConfig;
+import net.minecraft.resources.ResourceLocation;
 import net.minecraft.server.MinecraftServer;
 import org.apache.commons.io.FilenameUtils;
 
 import java.io.IOException;
-import java.io.InputStream;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
-import java.nio.file.StandardCopyOption;
+import java.util.Collections;
 import java.util.List;
-import java.util.Optional;
+import java.util.Map;
 
 import static me.drex.worldmanager.command.ImportCommand.*;
 import static net.minecraft.world.level.storage.LevelResource.LEVEL_DATA_FILE;
 
 public class RarArchiveExtractor implements ArchiveExtractor {
+    private Path root = Path.of(".");
+    private Archive archive;
+
     @Override
     public boolean supports(Path path) {
         String extension = FilenameUtils.getExtension(path.toString());
@@ -29,11 +32,10 @@ public class RarArchiveExtractor implements ArchiveExtractor {
     }
 
     @Override
-    public WorldConfig extract(Path rarPath, Path targetPath, MinecraftServer server) throws CommandSyntaxException, IOException {
-        Optional<WorldConfig> config = Optional.empty();
+    public Map<ResourceLocation, WorldConfig> open(Path rarPath, MinecraftServer server) throws IOException, CommandSyntaxException {
+        Map<ResourceLocation, WorldConfig> configs = Collections.emptyMap();
         try {
-            Archive archive = new Archive(Files.newInputStream(rarPath));
-            Path root = Path.of(".");
+            archive = new Archive(Files.newInputStream(rarPath));
             List<FileHeader> fileHeaders = archive.getFileHeaders();
             for (FileHeader fileHeader : fileHeaders) {
                 String fileName = fileHeader.getFileName().replace('\\', '/');
@@ -43,33 +45,11 @@ public class RarArchiveExtractor implements ArchiveExtractor {
                     if (parent != null) root = parent;
 
                     try (var is = archive.getInputStream(fileHeader)) {
-                        config = parseWorldConfig(is, server);
+                        configs = parseWorldConfig(is, server);
                     }
                     break;
                 }
             }
-            if (config.isEmpty()) throw MISSING_LEVEL_DAT.create(rarPath);
-
-            // Copy files
-            for (FileHeader fileHeader : fileHeaders) {
-                Path entryPath = Paths.get(fileHeader.getFileName().replace('\\', '/'));
-                Path relativize = root.relativize(entryPath);
-
-                String topDir = relativize.getName(0).toString();
-                if (!DIMENSION_PREFIXES.contains(topDir)) continue;
-
-                Path resolvedPath = targetPath.resolve(relativize);
-
-                if (fileHeader.isDirectory()) {
-                    Files.createDirectories(resolvedPath);
-                } else {
-                    Files.createDirectories(resolvedPath.getParent());
-                    try (InputStream inputStream = archive.getInputStream(fileHeader)) {
-                        Files.copy(inputStream, resolvedPath, StandardCopyOption.REPLACE_EXISTING);
-                    }
-                }
-            }
-            return config.get();
         } catch (RarException e) {
             if (e instanceof UnsupportedRarV5Exception) {
                 throw RAR5.create(rarPath);
@@ -77,5 +57,22 @@ public class RarArchiveExtractor implements ArchiveExtractor {
                 throw new IOException(e);
             }
         }
+        return configs;
+    }
+
+    @Override
+    public void extract(Path relativeSourcePath, Path targetRootPath, MinecraftServer server) throws IOException {
+        List<FileHeader> fileHeaders = archive.getFileHeaders();
+        for (FileHeader fileHeader : fileHeaders) {
+            Path entryPath = Paths.get(fileHeader.getFileName().replace('\\', '/'));
+            Path relativeToRoot = root.relativize(entryPath);
+
+            ArchiveExtractor.extractFile(relativeSourcePath, targetRootPath, relativeToRoot, fileHeader.isDirectory(), archive.getInputStream(fileHeader));
+        }
+    }
+
+    @Override
+    public void close() throws IOException {
+        archive.close();
     }
 }