+++ src/main/java/me/drex/worldmanager/extractor/TarGzArchiveExtractor.java
@@ -2,6 +2,7 @@ package me.drex.worldmanager.extractor;
 
 import com.mojang.brigadier.exceptions.CommandSyntaxException;
 import me.drex.worldmanager.save.WorldConfig;
+import net.minecraft.resources.ResourceLocation;
 import net.minecraft.server.MinecraftServer;
 import org.apache.commons.compress.archivers.tar.TarArchiveEntry;
 import org.apache.commons.compress.archivers.tar.TarArchiveInputStream;
@@ -14,13 +15,16 @@ import java.io.InputStream;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
-import java.nio.file.StandardCopyOption;
-import java.util.Optional;
+import java.util.Collections;
+import java.util.Map;
 
 import static me.drex.worldmanager.command.ImportCommand.*;
 import static net.minecraft.world.level.storage.LevelResource.LEVEL_DATA_FILE;
 
 public class TarGzArchiveExtractor implements ArchiveExtractor {
+    private Path root = Path.of(".");
+    private Path tarGzPath;
+
     @Override
     public boolean supports(Path path) {
         String name = FilenameUtils.getName(path.toString().toLowerCase());
@@ -28,15 +32,14 @@ public class TarGzArchiveExtractor implements ArchiveExtractor {
     }
 
     @Override
-    public WorldConfig extract(Path tarGzPath, Path targetPath, MinecraftServer server) throws CommandSyntaxException, IOException {
-        Optional<WorldConfig> config = Optional.empty();
-
+    public Map<ResourceLocation, WorldConfig> open(Path tarGzPath, MinecraftServer server) throws IOException, CommandSyntaxException {
+        this.tarGzPath = tarGzPath;
+        Map<ResourceLocation, WorldConfig> configs = Collections.emptyMap();
         try (InputStream fi = Files.newInputStream(tarGzPath);
              InputStream bi = new BufferedInputStream(fi);
              InputStream gzi = new GzipCompressorInputStream(bi);
              TarArchiveInputStream tarIn = new TarArchiveInputStream(gzi)) {
 
-            Path root = Path.of(".");
             TarArchiveEntry entry;
             while ((entry = tarIn.getNextEntry()) != null) {
                 Path entryPath = Paths.get(entry.getName());
@@ -44,38 +47,33 @@ public class TarGzArchiveExtractor implements ArchiveExtractor {
                     Path parent = entryPath.getParent();
                     if (parent != null) root = parent;
 
-                    config = parseWorldConfig(tarIn, server);
+                    configs = parseWorldConfig(tarIn, server);
                     break;
                 }
             }
+        }
+        return configs;
+    }
 
-            if (config.isEmpty()) throw MISSING_LEVEL_DAT.create(tarGzPath);
+    @Override
+    public void extract(Path relativeSourcePath, Path targetRootPath, MinecraftServer server) throws IOException {
+        try (InputStream fi2 = Files.newInputStream(tarGzPath);
+             InputStream bi2 = new BufferedInputStream(fi2);
+             InputStream gzi2 = new GzipCompressorInputStream(bi2);
+             TarArchiveInputStream tarIn2 = new TarArchiveInputStream(gzi2)) {
 
-            try (InputStream fi2 = Files.newInputStream(tarGzPath);
-                 InputStream bi2 = new BufferedInputStream(fi2);
-                 InputStream gzi2 = new GzipCompressorInputStream(bi2);
-                 TarArchiveInputStream tarIn2 = new TarArchiveInputStream(gzi2)) {
+            TarArchiveEntry entry;
+            while ((entry = tarIn2.getNextEntry()) != null) {
+                Path entryPath = Paths.get(entry.getName());
+                Path relativeToRoot = root.relativize(entryPath);
+                ArchiveExtractor.extractFile(relativeSourcePath, targetRootPath, relativeToRoot, entry.isDirectory(), tarIn2);
 
-                while ((entry = tarIn2.getNextEntry()) != null) {
-                    Path entryPath = Paths.get(entry.getName());
-                    Path relativize = root.relativize(entryPath);
-
-                    if (relativize.getNameCount() == 0) continue;
-                    String topDir = relativize.getName(0).toString();
-                    if (!DIMENSION_PREFIXES.contains(topDir)) continue;
-
-                    Path resolvedPath = targetPath.resolve(relativize);
-
-                    if (entry.isDirectory()) {
-                        Files.createDirectories(resolvedPath);
-                    } else {
-                        Files.createDirectories(resolvedPath.getParent());
-                        Files.copy(tarIn2, resolvedPath, StandardCopyOption.REPLACE_EXISTING);
-                    }
-                }
             }
         }
+    }
 
-        return config.get();
+    @Override
+    public void close() throws IOException {
+
     }
 }