+++ src/main/java/me/drex/worldmanager/extractor/ZipArchiveExtractor.java
@@ -1,19 +1,14 @@
 package me.drex.worldmanager.extractor;
 
-import com.mojang.brigadier.exceptions.CommandSyntaxException;
 import me.drex.worldmanager.save.WorldConfig;
+import net.minecraft.resources.ResourceLocation;
 import net.minecraft.server.MinecraftServer;
 import org.apache.commons.io.FilenameUtils;
 
 import java.io.IOException;
-import java.io.InputStream;
-import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
-import java.nio.file.StandardCopyOption;
-import java.util.Enumeration;
-import java.util.Iterator;
-import java.util.Optional;
+import java.util.*;
 import java.util.zip.ZipEntry;
 import java.util.zip.ZipFile;
 
@@ -21,6 +16,10 @@ import static me.drex.worldmanager.command.ImportCommand.*;
 import static net.minecraft.world.level.storage.LevelResource.LEVEL_DATA_FILE;
 
 public class ZipArchiveExtractor implements ArchiveExtractor {
+
+    private Path root = Path.of(".");
+    private ZipFile zipFile;
+
     @Override
     public boolean supports(Path path) {
         String extension = FilenameUtils.getExtension(path.toString());
@@ -28,48 +27,39 @@ public class ZipArchiveExtractor implements ArchiveExtractor {
     }
 
     @Override
-    public WorldConfig extract(Path zipPath, Path targetPath, MinecraftServer server) throws CommandSyntaxException, IOException {
-        Optional<WorldConfig> config = Optional.empty();
-        try (ZipFile zipFile = new ZipFile(zipPath.toFile())) {
-            Enumeration<? extends ZipEntry> entries = zipFile.entries();
-            Path root = Path.of(".");
-            while (entries.hasMoreElements()) {
-                ZipEntry zipEntry = entries.nextElement();
-                Path path = Path.of(zipEntry.getName());
-                if (path.getFileName().toString().equals(LEVEL_DATA_FILE.getId())) {
+    public Map<ResourceLocation, WorldConfig> open(Path zipPath, MinecraftServer server) throws IOException {
+        Map<ResourceLocation, WorldConfig> configs = Collections.emptyMap();
+        zipFile = new ZipFile(zipPath.toFile());
+        Enumeration<? extends ZipEntry> entries = zipFile.entries();
+        while (entries.hasMoreElements()) {
+            ZipEntry zipEntry = entries.nextElement();
+            Path path = Path.of(zipEntry.getName());
+            if (path.getFileName().toString().equals(LEVEL_DATA_FILE.getId())) {
+                try (var is = zipFile.getInputStream(zipEntry)) {
+                    configs = parseWorldConfig(is, server);
                     Path parent = path.getParent();
                     if (parent != null) root = parent;
-
-                    try (var is = zipFile.getInputStream(zipEntry)) {
-                        config = parseWorldConfig(is, server);
-                    }
-                    break;
                 }
-            }
-            if (config.isEmpty()) throw MISSING_LEVEL_DAT.create(zipPath);
-
-            // Copy files
-            Iterator<? extends ZipEntry> iterator = zipFile.entries().asIterator();
-            while (iterator.hasNext()) {
-                ZipEntry zipEntry = iterator.next();
-                Path entryPath = Paths.get(zipEntry.getName());
-                Path relativize = root.relativize(entryPath);
-
-                String topDir = relativize.getName(0).toString();
-                if (!DIMENSION_PREFIXES.contains(topDir)) continue;
-
-                Path resolvedPath = targetPath.resolve(relativize);
-
-                if (zipEntry.isDirectory()) {
-                    Files.createDirectories(resolvedPath);
-                } else {
-                    Files.createDirectories(resolvedPath.getParent());
-                    try (InputStream inputStream = zipFile.getInputStream(zipEntry)) {
-                        Files.copy(inputStream, resolvedPath, StandardCopyOption.REPLACE_EXISTING);
-                    }
-                }
+                break;
             }
         }
-        return config.get();
+        return configs;
+    }
+
+    @Override
+    public void extract(Path relativeSourcePath, Path targetRootPath, MinecraftServer server) throws IOException {
+        // Copy files
+        Iterator<? extends ZipEntry> iterator = zipFile.entries().asIterator();
+        while (iterator.hasNext()) {
+            ZipEntry zipEntry = iterator.next();
+            Path entryPath = Paths.get(zipEntry.getName());
+            Path relativeToRoot = root.relativize(entryPath);
+            ArchiveExtractor.extractFile(relativeSourcePath, targetRootPath, relativeToRoot, zipEntry.isDirectory(), zipFile.getInputStream(zipEntry));
+        }
+    }
+
+    @Override
+    public void close() throws IOException {
+        zipFile.close();
     }
 }